requirements.txt
attrs==19.3.0
Automat==0.8.0
biopython==1.78
blinker==1.4
certifi==2019.11.28
chardet==3.0.4
Click==7.0
cloud-init==20.4
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
cycler==0.10.0
dbus-python==1.2.16
decorator==4.4.2
Distance==0.1.3
distro==1.4.0
distro-info===0.23ubuntu1
entrypoints==0.3
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.5.0
incremental==16.10.1
Jinja2==2.10.1
joblib==1.0.1
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
kiwisolver==1.3.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
MarkupSafe==1.1.0
matplotlib==3.4.2
more-itertools==4.2.0
netifaces==0.10.4
networkx==2.5.1
numpy==1.20.3
oauthlib==3.1.0
pandas==1.2.4
pexpect==4.6.0
Pillow==8.2.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
PyGObject==3.36.0
pygraphviz==1.7
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
pymol==2.3.0
PyNaCl==1.3.0
PyOpenGL==3.1.0
pyOpenSSL==19.0.0
pyparsing==2.4.7
PyQt5==5.14.1
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.0+ubuntu0.20.4.4
python-dateutil==2.8.1
python-debian===0.1.36ubuntu1
python-louvain==0.15
pytz==2021.1
PyYAML==5.3.1
requests==2.22.0
requests-unixsocket==0.2.0
scikit-learn==0.24.2
scipy==1.6.3
seaborn==0.11.1
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
sip==4.19.21
six==1.14.0
sklearn==0.0
sos==4.0
ssh-import-id==5.10
systemd-python==234
threadpoolctl==2.1.0
Twisted==18.9.0
ubuntu-advantage-tools==20.3
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
wadllib==1.3.3
zipp==1.0.0
zope.interface==4.7.1
